port(GatherBuddyReborn): TC API12 source (based on TC_forward @ e7c22014)

+++ OtterGui/Classes/LowerString.cs
@@ -1,4 +1,4 @@
-using ImGuiNET;
+using Dalamud.Bindings.ImGui;
 using Newtonsoft.Json;
 
 namespace OtterGui.Classes;
@@ -71,11 +71,11 @@ public readonly struct LowerString : IEquatable<LowerString>, IComparable<LowerS
         => new(s);
 
     // Create an ImGui text input box with hint that converts the string to a LowerString.
-    public static bool InputWithHint(string label, string hint, ref LowerString s, uint maxLength = 128,
+    public static bool InputWithHint(string label, string hint, ref LowerString s, int maxLength = 128,
         ImGuiInputTextFlags flags = ImGuiInputTextFlags.None)
     {
         var tmp = s.Text;
-        if (!ImGui.InputTextWithHint(label, hint, ref tmp, maxLength, flags) || tmp == s.Text)
+        if (!ImGui.InputTextWithHint(label, hint, ref tmp, (int)maxLength, flags) || tmp == s.Text)
             return false;
 
         s = new LowerString(tmp);
